Disable C# Code Actions in @functions blocks
Fixes: https://github.com/dotnet/aspnetcore/issues/26715
Revert PR: https://github.com/dotnet/aspnetcore/issues/26867
