+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/CodeActions/CSharp/DefaultCSharpCodeActionProvider.cs
@@ -51,6 +51,13 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.CodeActions
                 return EmptyResult;
             }
 
+            // Disable multi-line code actions in @functions block
+            // Will be removed once https://github.com/dotnet/aspnetcore/issues/26501 is unblocked.
+            if (InFunctionsBlock(context))
+            {
+                return EmptyResult;
+            }
+
             var results = codeActions.Where(codeAction =>
                 StringMatchCodeActions.Contains(codeAction.Title) ||
                 RegexMatchCodeActions.Any(pattern => pattern.Match(codeAction.Title).Success)
