+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/CodeActions/CSharp/ImplementInterfaceAbstractClassCodeActionProvider.cs
@@ -57,6 +57,13 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.CodeActions
                 return EmptyResult;
             }
 
+            // Disable multi-line code actions in @functions block
+            // Will be removed once https://github.com/dotnet/aspnetcore/issues/26501 is unblocked.
+            if (InFunctionsBlock(context))
+            {
+                return EmptyResult;
+            }
+
             var diagnostics = context.Request.Context.Diagnostics.Where(diagnostic =>
                     diagnostic.Severity == DiagnosticSeverity.Error &&
                     diagnostic.Code?.IsString == true)
