+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/CodeActions/CSharp/DefaultCSharpCodeActionResolverTest.cs
@@ -1,120 +1,276 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Test.Common;
-using Microsoft.AspNetCore.Razor.LanguageServer.CodeActions.Models;
-using Xunit;
-using System.Linq;
-using OmniSharp.Extensions.LanguageServer.Protocol.Models;
-using Moq;
-using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.AspNetCore.Razor.Language;
+using Microsoft.AspNetCore.Razor.LanguageServer.CodeActions.Models;
+using Microsoft.AspNetCore.Razor.LanguageServer.Common;
+using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.CodeAnalysis.Text;
+using Moq;
+using OmniSharp.Extensions.LanguageServer.Protocol.Models;
+using Xunit;
+using OmniSharp.Extensions.LanguageServer.Protocol.Server;
+using OmniSharp.Extensions.JsonRpc;
+using System.Threading;
+using Microsoft.AspNetCore.Razor.LanguageServer.Formatting;
+using Newtonsoft.Json.Linq;
+using Microsoft.AspNetCore.Razor.LanguageServer.ProjectSystem;
+using Microsoft.AspNetCore.Mvc.Razor.Extensions;
+using System.Linq;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.CodeActions
 {
-    public class DefaultCSharpCodeActionProviderTest : LanguageServerTestBase
+    public class DefaultCSharpCodeActionResolverTest : LanguageServerTestBase
     {
-        private static readonly RazorCodeAction[] SupportedCodeActions = new RazorCodeAction[]
+        private static readonly RazorCodeAction DefaultResolvedCodeAction = new RazorCodeAction()
         {
-            new RazorCodeAction()
+            Title = "ResolvedCodeAction",
+            Data = new object(),
+            Edit = new WorkspaceEdit()
             {
-                Title = "Generate Equals and GetHashCode"
-            },
-            new RazorCodeAction()
-            {
-                Title = "Add null check"
-            },
-            new RazorCodeAction()
-            {
-                Title = "Add null checks for all parameters"
-            },
-            new RazorCodeAction()
-            {
-                Title = "Add null checks for all parameters"
-            },
-            new RazorCodeAction()
-            {
-                Title = "Generate constructor 'Counter(int)'"
+                DocumentChanges = new Container<WorkspaceEditDocumentChange>(
+                    new WorkspaceEditDocumentChange(
+                        new TextDocumentEdit()
+                        {
+                            Edits = new TextEditContainer(
+                                new TextEdit()
+                                {
+                                    NewText = "Generated C# Based Edit"
+                                }
+                            )
+                        }
+                    ))
             }
         };
 
-        [Fact]
-        public async Task ProvideAsync_ValidCodeActions_ReturnsProvidedCodeAction()
+        private static readonly TextEdit[] DefaultFormattedEdits = new TextEdit[]
         {
-            // Arrange
-            var provider = new DefaultCSharpCodeActionProvider();
-            var context = CreateCodeActionContext(supportsCodeActionResolve: true);
-
-            // Act
-            var providedCodeActions = await provider.ProvideAsync(context, SupportedCodeActions, default);
-
-            // Assert
-            Assert.Equal(SupportedCodeActions.Length, providedCodeActions.Count);
-
-            for (var i = 0; i < SupportedCodeActions.Length; i++)
+            new TextEdit()
             {
-                Assert.Equal(SupportedCodeActions[i].Title, providedCodeActions.ElementAt(i).Title);
+                NewText = "Remapped & Formatted Edit"
             }
-        }
+        };
+
+        private static readonly RazorCodeAction DefaultUnresolvedCodeAction = new RazorCodeAction()
+        {
+            Title = "Unresolved Code Action"
+        };
 
         [Fact]
-        public async Task ProvideAsync_SupportsCodeActionResolveFalse_ValidCodeActions_ReturnsEmpty()
+        public async Task ResolveAsync_ReturnsResolvedCodeAction()
         {
             // Arrange
-            var provider = new DefaultCSharpCodeActionProvider();
-            var context = CreateCodeActionContext(supportsCodeActionResolve: false);
+            CreateCodeActionResolver(out var codeActionParams, out var csharpCodeActionResolver);
 
             // Act
-            var providedCodeActions = await provider.ProvideAsync(context, SupportedCodeActions, default);
+            var returnedCodeAction = await csharpCodeActionResolver.ResolveAsync(codeActionParams, DefaultUnresolvedCodeAction, default);
 
             // Assert
-            Assert.Empty(providedCodeActions);
+            Assert.Equal(DefaultResolvedCodeAction.Title, returnedCodeAction.Title);
+            Assert.Equal(DefaultResolvedCodeAction.Data, returnedCodeAction.Data);
+            var returnedEdits = Assert.Single(returnedCodeAction.Edit.DocumentChanges);
+            Assert.True(returnedEdits.IsTextDocumentEdit);
+            var returnedTextDocumentEdit = Assert.Single(returnedEdits.TextDocumentEdit.Edits);
+            Assert.Equal(DefaultFormattedEdits.First(), returnedTextDocumentEdit);
         }
 
         [Fact]
-        public async Task ProvideAsync_InvalidCodeActions_ReturnsNoCodeActions()
+        public async Task ResolveAsync_NoDocumentChanges_ReturnsOriginalCodeAction()
         {
             // Arrange
-            var provider = new DefaultCSharpCodeActionProvider();
-            var context = CreateCodeActionContext(supportsCodeActionResolve: true);
-
-            var codeActions = new RazorCodeAction[]
+            var resolvedCodeAction = new RazorCodeAction()
             {
-                new RazorCodeAction()
+                Title = "ResolvedCodeAction",
+                Data = new object(),
+                Edit = new WorkspaceEdit()
                 {
-                    Title = "Do something not really supported in razor"
-                },
-                new RazorCodeAction()
-                {
-                    // Invalid regex pattern shouldn't match
-                    Title = "Generate constructor 'Counter(int)' xyz"
+                    DocumentChanges = null
                 }
             };
 
+            var languageServer = CreateLanguageServer(resolvedCodeAction);
+
+            CreateCodeActionResolver(out var codeActionParams, out var csharpCodeActionResolver, languageServer: languageServer);
+
             // Act
-            var providedCodeActions = await provider.ProvideAsync(context, codeActions, default);
+            var returnedCodeAction = await csharpCodeActionResolver.ResolveAsync(codeActionParams, DefaultUnresolvedCodeAction, default);
 
             // Assert
-            Assert.Empty(providedCodeActions);
+            Assert.Equal(DefaultUnresolvedCodeAction.Title, returnedCodeAction.Title);
         }
 
-        private static RazorCodeActionContext CreateCodeActionContext(bool supportsCodeActionResolve)
+        [Fact]
+        public async Task ResolveAsync_MultipleDocumentChanges_ReturnsOriginalCodeAction()
         {
-            var codeActionParams = new CodeActionParams()
+            // Arrange
+            var resolvedCodeAction = new RazorCodeAction()
             {
-                TextDocument = new TextDocumentIdentifier("c:/path/to/file.razor")
+                Title = "ResolvedCodeAction",
+                Data = new object(),
+                Edit = new WorkspaceEdit()
+                {
+                    DocumentChanges = new Container<WorkspaceEditDocumentChange>(
+                        new WorkspaceEditDocumentChange(
+                            new TextDocumentEdit()
+                            {
+                                Edits = new TextEditContainer(
+                                    new TextEdit()
+                                    {
+                                        NewText = "1. Generated C# Based Edit"
+                                    }
+                                )
+                            }
+                        ),
+                        new WorkspaceEditDocumentChange(
+                            new TextDocumentEdit()
+                            {
+                                Edits = new TextEditContainer(
+                                    new TextEdit()
+                                    {
+                                        NewText = "2. Generated C# Based Edit"
+                                    }
+                                )
+                            }
+                        ))
+                }
             };
-            var context = new RazorCodeActionContext(
-                codeActionParams,
-                Mock.Of<DocumentSnapshot>(),
-                Mock.Of<RazorCodeDocument>(),
-                new SourceLocation(),
-                SourceText.From(string.Empty),
-                supportsFileCreation: false,
-                supportsCodeActionResolve);
-            return context;
+
+            var languageServer = CreateLanguageServer(resolvedCodeAction);
+
+            CreateCodeActionResolver(out var codeActionParams, out var csharpCodeActionResolver, languageServer: languageServer);
+
+            // Act
+            var returnedCodeAction = await csharpCodeActionResolver.ResolveAsync(codeActionParams, DefaultUnresolvedCodeAction, default);
+
+            // Assert
+            Assert.Equal(DefaultUnresolvedCodeAction.Title, returnedCodeAction.Title);
+        }
+
+        [Fact]
+        public async Task ResolveAsync_NonTextDocumentEdit_ReturnsOriginalCodeAction()
+        {
+            // Arrange
+            var resolvedCodeAction = new RazorCodeAction()
+            {
+                Title = "ResolvedCodeAction",
+                Data = new object(),
+                Edit = new WorkspaceEdit()
+                {
+                    DocumentChanges = new Container<WorkspaceEditDocumentChange>(
+                        new WorkspaceEditDocumentChange(
+                            new CreateFile()
+                            {
+                                Uri = new Uri("c:/some/uri.razor")
+                            }
+                        ))
+                }
+            };
+
+            var languageServer = CreateLanguageServer(resolvedCodeAction);
+
+            CreateCodeActionResolver(out var codeActionParams, out var csharpCodeActionResolver, languageServer: languageServer);
+
+            // Act
+            var returnedCodeAction = await csharpCodeActionResolver.ResolveAsync(codeActionParams, DefaultUnresolvedCodeAction, default);
+
+            // Assert
+            Assert.Equal(DefaultUnresolvedCodeAction.Title, returnedCodeAction.Title);
+        }
+
+        private void CreateCodeActionResolver(
+            out CSharpCodeActionParams codeActionParams,
+            out DefaultCSharpCodeActionResolver csharpCodeActionResolver,
+            IClientLanguageServer languageServer = null,
+            DocumentVersionCache documentVersionCache = null,
+            RazorFormattingService razorFormattingService = null)
+        {
+            var documentPath = "c:/Test.razor";
+            var documentUri = new Uri(documentPath);
+            var contents = string.Empty;
+            var codeDocument = CreateCodeDocument(contents, documentPath);
+
+            codeActionParams = new CSharpCodeActionParams()
+            {
+                Data = new JObject(),
+                RazorFileUri = documentUri
+            };
+
+            languageServer ??= CreateLanguageServer();
+            documentVersionCache ??= CreateDocumentVersionCache();
+            razorFormattingService ??= CreateRazorFormattingService(documentUri);
+
+            csharpCodeActionResolver = new DefaultCSharpCodeActionResolver(
+                new DefaultForegroundDispatcher(),
+                CreateDocumentResolver(documentPath, codeDocument),
+                languageServer,
+                razorFormattingService,
+                documentVersionCache);
+        }
+
+        private static RazorFormattingService CreateRazorFormattingService(Uri documentUri)
+        {
+            var razorFormattingService = Mock.Of<RazorFormattingService>(
+                            rfs => rfs.ApplyFormattedEditsAsync(
+                                documentUri,
+                                It.IsAny<DocumentSnapshot>(),
+                                RazorLanguageKind.CSharp,
+                                It.IsAny<TextEdit[]>(),
+                                It.IsAny<FormattingOptions>(),
+                                It.IsAny<CancellationToken>(),
+                                /*bypassValidationPasses:*/ true) == Task.FromResult(DefaultFormattedEdits));
+            return razorFormattingService;
+        }
+
+        private static DocumentVersionCache CreateDocumentVersionCache()
+        {
+            int? documentVersion = 2;
+            var documentVersionCache = Mock.Of<DocumentVersionCache>(dvc => dvc.TryGetDocumentVersion(It.IsAny<DocumentSnapshot>(), out documentVersion) == true);
+            return documentVersionCache;
+        }
+
+        private IClientLanguageServer CreateLanguageServer(RazorCodeAction resolvedCodeAction = null)
+        {
+            var responseRouterReturns = new Mock<IResponseRouterReturns>(MockBehavior.Strict);
+            responseRouterReturns
+                .Setup(l => l.Returning<RazorCodeAction>(It.IsAny<CancellationToken>()))
+                .Returns(Task.FromResult(resolvedCodeAction ?? DefaultResolvedCodeAction));
+
+            var languageServer = new Mock<IClientLanguageServer>(MockBehavior.Strict);
+            languageServer
+                .Setup(l => l.SendRequest(LanguageServerConstants.RazorResolveCodeActionsEndpoint, It.IsAny<RazorCodeAction>()))
+                .Returns(responseRouterReturns.Object);
+
+            return languageServer.Object;
+        }
+
+        private static DocumentResolver CreateDocumentResolver(string documentPath, RazorCodeDocument codeDocument)
+        {
+            var sourceTextChars = new char[codeDocument.Source.Length];
+            codeDocument.Source.CopyTo(0, sourceTextChars, 0, codeDocument.Source.Length);
+            var sourceText = SourceText.From(new string(sourceTextChars));
+            var documentSnapshot = Mock.Of<DocumentSnapshot>(document =>
+                document.GetGeneratedOutputAsync() == Task.FromResult(codeDocument) &&
+                document.GetTextAsync() == Task.FromResult(sourceText));
+            var documentResolver = new Mock<DocumentResolver>();
+            documentResolver
+                .Setup(resolver => resolver.TryResolveDocument(documentPath, out documentSnapshot))
+                .Returns(true);
+            return documentResolver.Object;
+        }
+
+        private static RazorCodeDocument CreateCodeDocument(string text, string documentPath)
+        {
+            var projectItem = new TestRazorProjectItem(documentPath) { Content = text };
+            var projectEngine = RazorProjectEngine.Create(RazorConfiguration.Default, TestRazorProjectFileSystem.Empty, (builder) =>
+            {
+                PageDirective.Register(builder);
+            });
+            var codeDocument = projectEngine.Process(projectItem);
+            codeDocument.SetFileKind(FileKinds.Component);
+            return codeDocument;
         }
     }
 }
