+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/CodeActions/CSharp/ImplementInterfaceAbstractClassCodeActionProviderTest.cs
@@ -216,7 +216,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.CodeActions
         }
 
         [Fact]
-        public async Task Handle_ValidDiagnostic_ValidCodeAction_ReturnsCodeActions()
+        public async Task Handle_ValidDiagnostic_ValidCodeAction_CodeBlock_ReturnsCodeActions()
         {
             // Arrange
             var documentPath = "c:/Test.razor";
@@ -260,7 +260,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.CodeActions
                 }
             };
 
-            var location = new SourceLocation(0, -1, -1);
+            var location = new SourceLocation(8, -1, -1);
             var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(8, 4));
             context.CodeDocument.SetFileKind(FileKinds.Legacy);
 
@@ -300,6 +300,74 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.CodeActions
                     Assert.IsType<CSharpCodeActionParams>(resolutionParams.Data);
                 }
             );
+        }
+
+        [Fact]
+        public async Task Handle_ValidDiagnostic_ValidCodeAction_FunctionsBlock_ReturnsEmpty()
+        {
+            // Arrange
+            var documentPath = "c:/Test.razor";
+            var contents = "@functions { Path; }";
+            var request = new CodeActionParams()
+            {
+                TextDocument = new TextDocumentIdentifier(new Uri(documentPath)),
+                Range = new Range(),
+                Context = new CodeActionContext()
+                {
+                    Diagnostics = new Container<Diagnostic>(
+                        new Diagnostic()
+                        {
+                            Severity = DiagnosticSeverity.Error,
+                            Code = new DiagnosticCode("CS0132")
+                        },
+                        new Diagnostic()
+                        {
+                            Severity = DiagnosticSeverity.Error,
+                            Code = new DiagnosticCode("CS0534"),
+                            Range = new Range(
+                                new Position(0, 13),
+                                new Position(0, 17)
+                            )
+                        },
+                        new Diagnostic()
+                        {
+                            Severity = DiagnosticSeverity.Error,
+                            Code = new DiagnosticCode("CS0535"),
+                            Range = new Range(
+                                new Position(0, 13),
+                                new Position(0, 17)
+                            )
+                        },
+                        new Diagnostic()
+                        {
+                            Severity = DiagnosticSeverity.Error,
+                            Code = new DiagnosticCode("CS0183")
+                        }
+                    )
+                }
+            };
+
+            var location = new SourceLocation(13, -1, -1);
+            var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(13, 4));
+            context.CodeDocument.SetFileKind(FileKinds.Legacy);
+
+            var provider = new ImplementInterfaceAbstractClassCodeActionProvider();
+            var csharpCodeActions = new[] {
+                new RazorCodeAction()
+                {
+                    Title = "Implement abstract class"
+                },
+                new RazorCodeAction()
+                {
+                    Title = "Implement interface"
+                }
+            };
+
+            // Act
+            var results = await provider.ProvideAsync(context, csharpCodeActions, default);
+
+            // Assert
+            Assert.Empty(results);
         }
 
         [Fact]
